Extract data pipeline to daemon
