+++ app.py
@@ -1,8 +1,6 @@
 from __future__ import annotations
 
 import os
-from datetime import datetime
-from apscheduler.schedulers.background import BackgroundScheduler
 from flask import Flask
 from dotenv import load_dotenv
 
@@ -20,7 +18,6 @@ from routes.admin import bp as admin_bp
 from routes.api_datasets import bp as api_datasets_bp
 from routes.api_projects import bp as api_projects_bp
 from routes.api_dashboards import bp as api_dashboards_bp
-from services.data_ingestion import cache_series, fetch_remote_series
 from ws import init_app as init_ws, socketio
 from models.db import Base, engine
 
@@ -47,22 +44,6 @@ app.register_blueprint(api_dashboards_bp)
 init_ws(app)
 
 load_dotenv()
-
-SOURCE_URL = os.environ.get("SERIES_SOURCE", "https://example.com/api")
-SYMBOL = os.environ.get("SERIES_SYMBOL", "demo")
-START_DATE = os.environ.get("SERIES_START", "2000-01-01")
-END_DATE = os.environ.get("SERIES_END", datetime.utcnow().strftime("%Y-%m-%d"))
-
-scheduler = BackgroundScheduler()
-
-
-def fetch_and_cache() -> None:
-    df = fetch_remote_series(SOURCE_URL, SYMBOL, START_DATE, END_DATE)
-    cache_series(SYMBOL, df)
-
-
-scheduler.add_job(fetch_and_cache, "cron", hour=0, minute=0)
-scheduler.start()
 
 
 @app.route("/")
